fix(physics2d): fix StaticCollider shapes cleared on reuse
Move _shapes = null from _removeShapeAndDestroyData to _onDestroy to prevent
clearing shapes array when the collider is reused. Fixes issue where reused
StaticCollider loses shape data during shape removal operations.

Also includes migration of Rand class from d3/math/Rand.ts to maths/Rand.ts
and updates all particle system imports to use the new location.
