+++ src/layaAir/laya/particle/d3/module/shape/HemisphereShape.ts
@@ -1,5 +1,5 @@
 import { BoundBox } from "../../../../d3/math/BoundBox";
-import { Rand } from "../../../../d3/math/Rand";
+import { Rand } from "../../../../maths/Rand";
 import { Vector3 } from "../../../../maths/Vector3";
 import { BaseShape, ParticleSystemShapeType } from "./BaseShape";
 import { ShapeUtils } from "./ShapeUtils";
