+++ src/layaAir/laya/particle/d3/module/shape/ShapeUtils.ts
@@ -1,4 +1,4 @@
-import { Rand } from "../../../../d3/math/Rand";
+import { Rand } from "../../../../maths/Rand";
 import { Vector2 } from "../../../../maths/Vector2";
 import { Vector3 } from "../../../../maths/Vector3";
